pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "face-match"
version = "0.2.0"
description = "Búsqueda de coincidencias faciales ultra rápida (OpenCV YuNet + SFace + FAISS)."
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [{ name = "MOCA" }]
keywords = ["face-recognition", "opencv", "faiss", "biometrics", "face-search", "sface"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Intended Audience :: System Administrators",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Image Processing",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Typing :: Typed",
]
dependencies = [
  "opencv-python>=4.8.0",
  "numpy>=1.24.0",
  "faiss-cpu>=1.7.4",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
]

[project.urls]
"Homepage" = "https://github.com/Moca9801/tracking-face"
"Bug Tracker" = "https://github.com/Moca9801/tracking-face/issues"

[project.scripts]
face-match = "face_match.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
face_match = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"
